#254 fix exception when user node points to nodes with the same app name
An exception is thrown when drawing the server map in situations where
users call multiple nodes with the same application name (but different
service types). This was caused because it was not possible to have a
single user node pointing to more than 1 destination node.

The user nodes are uniquely identified by their destination node's
application name only and this resulted in only a single user node being
created when there were multiple nodes with the same application name.

This commit fixes the issue by taking the destination node's service
type into account when creating user nodes.