#926 Allow applications without transaction to be shown in the servermap
The nodes in the server map were drawn based on whether or not their
agents have response data. Thus it was not possible to tell if an
application (more specifically its agents) was dead, or if it simply did
not have any transaction.

Server map now renders a node if it has running agents (based on the
higher bound of the queried time range).