[#5053] Allow plugins to specify compiler version
By allowing plugins to specify and be built using different jdk versions, we need a way to define trace metadata without loading TraceMetadataProvider implementations from plugins.
This is because the service loader will throw class version error when the java version running collector/web is lower than the jdk version used to compile the plugins.

To circumvent this, the commit adds a way to define trace metadata as a yaml file, so that type definitions may be parsed without loading them via the service loader.